add update/delete post upload image

+++ app/Http/Controllers/PostController.php
@@ -4,10 +4,22 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Storage;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth', except: ['index', 'show']),
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -33,17 +45,27 @@ class PostController extends Controller
     public function store(Request $request)
     {
         // Validate
-        $fields = $request->validate([
+        $request->validate([
             'title' => ['required', 'max:255'],
             'body' => ['required'],
+            'image' => ['nullable', 'file', 'max:3000', 'mimes:webp,png,jpg'],
         ]);
 
+        //Store image if it exists
+        $path = null;
+        if ($request->hasFile('image')) {
+            $path = Storage::disk('public')->put('posts_images', $request->image);
+        }
+
         //Create Post
-        Auth::user()->posts()->create($fields);
+        Auth::user()->posts()->create([
+            'title' => $request->title,
+            'body' => $request->body,
+            'image' => $path,
+        ]);
 
         // Redirect to dashboard
-        return back()->with('success', 'Successfully created');
-        dd('ok');
+        return back()->with('success', 'Successfully created!');
     }
 
     /**
@@ -59,7 +81,9 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        //
+        Gate::authorize('modify', $post);
+
+        return view('posts.edit', ['post' => $post]);
     }
 
     /**
@@ -67,7 +91,34 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        //
+        //Authorizing the action
+        Gate::authorize('modify', $post);
+
+        // Validate
+        $request->validate([
+            'title' => ['required', 'max:255'],
+            'body' => ['required'],
+            'image' => ['nullable', 'file', 'max:3000', 'mimes:webp,png,jpg'],
+        ]);
+
+        //Store image if it exists
+        $path = $post->image ?? null;
+        if ($request->hasFile('image')) {
+            if ($post->image) {
+                Storage::disk('public')->delete($post->image);
+            }
+            $path = Storage::disk('public')->put('posts_images', $request->image);
+        }
+
+        //Update Post
+        $post->update([
+            'title' => $request->title,
+            'body' => $request->body,
+            'image' => $path,
+        ]);
+
+        // Redirect to dashboard
+        return redirect()->route('dashboard')->with('success', 'Successfully Updated!');
     }
 
     /**
@@ -75,6 +126,14 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        //Authorizing the action
+        Gate::authorize('modify', $post);
+
+        //Delete image if exists
+        if ($post->image) {
+            Storage::disk('public')->delete($post->image);
+        }
+
         //Delete the post
         $post->delete();
